Refactorisation du contrôleur, amélioration de la gestion des entités, et petites modifications dans la partie View et Model / by FullStack
