+++ app/Controllers/AuthentifController.php
@@ -56,69 +56,41 @@ class AuthentifController extends BaseController{
     }
 
     // Processus de Login
-    public function SignIn_Process() {
-        $accountmodel = new AccountModel();
-    
-        // Récupérer les données du formulaire
-        $email = $this->request->getPost('logemail');
-        $password = $this->request->getPost('logpass');
-    
-        // Chercher l'utilisateur dans la table accounts
-        $User = $accountmodel->where('AcademicEmail', $email)->first();
-    
-        // Vérifier si l'utilisateur existe
-        if (!$User) {
-            return redirect()->back()->withInput()->with('general_error', 'Email not found. Please check your email.');
-        }
-    
-        // Vérifier si le mot de passe est haché
-        if (password_verify($password, $User['Password'])) {
-            // Authentification réussie
-            session()->set('user', [
-                'UserID' => $User['UserID'],
-                'AcademicEmail' => $User['AcademicEmail'],
-                'AccountID' => $User['AccountID'],
-                'RoleID' => $User['RoleID'],
-            ]);
-    
-            // Rediriger vers le dashboard approprié
-            if ($User['RoleID'] == 1) {
+    public function SignIn_Process()
+{
+    $accountmodel = new AccountModel();
+
+    // Récupérer les données du formulaire
+    $email = $this->request->getPost('logemail');
+    $password = $this->request->getPost('logpass');
+
+    // Chercher l'utilisateur dans la table accounts
+    $User = $accountmodel->where('AcademicEmail', $email)->first();
+
+    // Vérifier si l'utilisateur existe
+    if (!$User) {
+        return redirect()->back()->withInput()->with('general_error', 'Email not found. Please check your email.');
+    }
+
+    // Vérifier si le mot de passe est correct
+    if (password_verify($password, $User->Password)) {
+        // Authentification réussie
+        session()->set('user', $User->toArray());
+
+        // Rediriger vers le tableau de bord approprié
+        switch ($User->RoleID) {
+            case 1:
                 return redirect()->to('/StudentDashboard');
-            } elseif ($User['RoleID'] == 2) {
+            case 2:
                 return redirect()->to('/ProfDashboard');
-            } else {
-                return redirect()->to('/login')->with('error', 'ERROR: Ops! Try again.');
-            }
-        } else {
-            // Si le mot de passe n'est pas haché (ancien utilisateur)
-            if ($User['Password'] === $password) {
-                // Authentification réussie avec mot de passe non haché
-                // Hacher le mot de passe et mettre à jour dans la base
-                $hashedPassword = password_hash($password, PASSWORD_DEFAULT);
-                $accountmodel->update($User['AccountID'], ['Password' => $hashedPassword]);
-    
-                // Créer une session utilisateur
-                session()->set('user', [
-                    'UserID' => $User['UserID'],
-                    'AcademicEmail' => $User['AcademicEmail'],
-                    'AccountID' => $User['AccountID'],
-                    'RoleID' => $User['RoleID'],
-                ]);
-    
-                // Rediriger vers le dashboard approprié
-                if ($User['RoleID'] == 1) {
-                    return redirect()->to('/StudentDashboard');
-                } elseif ($User['RoleID'] == 2) {
-                    return redirect()->to('/ProfDashboard');
-                } else {
-                    return redirect()->to('/login')->with('error', 'ERROR: Ops! Try again.');
-                }
-            } else {
-                // Si le mot de passe est incorrect
-                return redirect()->back()->withInput()->with('general_error', 'Password incorrect! Please try again.');
-            }
+            default:
+                return redirect()->to('/login')->with('error', 'Role not recognized. Try again.');
         }
     }
+
+    // Mot de passe incorrect
+    return redirect()->back()->withInput()->with('general_error', 'Password incorrect! Please try again.');
+}
     
     
     //Processus de Logout
