--- app/Controllers/Home.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Controllers;
-
-class Home extends BaseController
-{
-    public function index(): string
-    {
-        return view('welcome_message');
-    }
-}
