--- app/Controllers/Register.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Controllers;
-
-class Register extends BaseController
-{
-    public function index()
-    {
-        return view('login'); // Charge la vue 'register.php'
-}
-}
