+++ app/Controllers/StudentDashboard_Controller.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Controllers;
+
+class StudentDashboard_Controller extends BaseController{
+        // Dashboard pour Student
+        public function StudentDashboard()
+        {
+            // Récupérer les informations spécifiques à l'étudiant
+            return view('student_dashboard');
+        }
+}
