--- app/Controllers/loginController.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Controllers;
-
-class loginController extends BaseController
-{
-    public function index()
-    {
-        return view('login'); // Remplacez par le nom de votre fichier vue
-    }
-}
