+++ app/Entities/Roles.php
@@ -1,5 +1,5 @@
 <?php
-namespace App\Models\Entities;
+namespace App\Entities;
 
 use CodeIgniter\Entity\Entity;
