+++ writable/session/ci_session2633u5u9f6scrgbgdq87r09oej51694f
@@ -1,0 +1,1 @@
+__ci_last_regenerate|i:1736715595;_ci_previous_url|s:46:"http://localhost/GestionDeExamens/public/login";
