+++ writable/session/ci_session6510fn662u4tbnr1rct1om55gd3jb9u6
@@ -1,0 +1,1 @@
+__ci_last_regenerate|i:1736717695;_ci_previous_url|s:48:"http://localhost/GestionDeExamens/public/Modules";user|a:5:{s:9:"AccountID";s:2:"11";s:6:"UserID";s:2:"10";s:6:"RoleID";s:1:"2";s:13:"AcademicEmail";s:27:"Adil.Rachidi@prof.uiz.ac.ma";s:8:"Password";s:60:"$2y$10$VDVNL1y9QajOeJaxm4JFV.94GLyKbnB4tf8ncy4SpM59s.RU96Xim";}
