+++ writable/session/ci_session9b368t84mfft70nbrbb5gjodqjqm0q20
@@ -1,0 +1,1 @@
+__ci_last_regenerate|i:1736716293;_ci_previous_url|s:46:"http://localhost/GestionDeExamens/public/login";user|a:5:{s:9:"AccountID";s:1:"7";s:6:"UserID";s:1:"9";s:6:"RoleID";s:1:"2";s:13:"AcademicEmail";s:27:"Meskin.Ahmed@prof.uiz.ac.ma";s:8:"Password";s:60:"$2y$10$uEHhu3Xw3Otps1rUIrIJR.xWK/InQz8.eBl9y9sjNidSt6Vg0aphO";}
