+++ writable/session/ci_sessioni8tmrbn9amv6hskqafumuqahj7hjekuq
@@ -1,0 +1,1 @@
+__ci_last_regenerate|i:1736717945;_ci_previous_url|s:46:"http://localhost/GestionDeExamens/public/login";
